release: support tms cloud.

+++ src/main/java/ru/testit/client/model/CreateWorkItemApiModel.java
@@ -82,7 +82,7 @@ public class CreateWorkItemApiModel {
 
   public static final String JSON_PROPERTY_DURATION = "duration";
   @jakarta.annotation.Nonnull
-  private Integer duration;
+  private Long duration;
 
   public static final String JSON_PROPERTY_STATE = "state";
   @jakarta.annotation.Nonnull
@@ -209,7 +209,7 @@ public class CreateWorkItemApiModel {
   }
 
 
-  public CreateWorkItemApiModel duration(@jakarta.annotation.Nonnull Integer duration) {
+  public CreateWorkItemApiModel duration(@jakarta.annotation.Nonnull Long duration) {
     this.duration = duration;
     return this;
   }
@@ -224,14 +224,14 @@ public class CreateWorkItemApiModel {
   @JsonProperty(JSON_PROPERTY_DURATION)
   @JsonInclude(value = JsonInclude.Include.ALWAYS)
 
-  public Integer getDuration() {
+  public Long getDuration() {
     return duration;
   }
 
 
   @JsonProperty(JSON_PROPERTY_DURATION)
   @JsonInclude(value = JsonInclude.Include.ALWAYS)
-  public void setDuration(@jakarta.annotation.Nonnull Integer duration) {
+  public void setDuration(@jakarta.annotation.Nonnull Long duration) {
     this.duration = duration;
   }
